backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance>=0.2.31
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
statsmodels>=0.14.0
anthropic>=0.39.0
apscheduler>=3.10.0
requests>=2.31.0
feedparser>=6.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
